workflow: publish a multi-platform docker image
I updated the Dockerfile to use a multi-stage build. By building the binary in the dockerfile we don't need to worry about how it will run on different architectures. However, I noticed a significant increase in the time it takes to run the workflow. This will likely need revisited in the future.